--- lab05/tienda/models.py
@@ -1,32 +1,0 @@
-from django.db import models
-from datetime import datetime   
-# Create your models here.
-
-class Categoria(models.Model):
-    nombre = models.CharField(max_length=200)
-    pub_dat = models.DateTimeField('date published')
-
-    def __str__(self):
-        return self.nombre
-
-class Producto(models.Model):
-    categoria = models.ForeignKey(Categoria, on_delete=models.CASCADE)
-    nombre = models.CharField(max_length=200)
-    precio = models.DecimalField(max_digits=6, decimal_places=2)
-    stock = models.IntegerField(default=0)
-    pub_date = models.DateTimeField('date published')
-
-    def __str__(self):
-        return self.nombre
-
-class Cliente(models.Model):
-    nombre = models.CharField(max_length=200)
-    apellidos = models.CharField(max_length=200)
-    dni = models.IntegerField(unique=True)
-    telefono = models.IntegerField(default=0)
-    email = models.CharField(max_length=200)
-    fecha_nacimiento = models.DateTimeField('fecha nacimiento')
-    pub_date = models.DateTimeField('date published',default=datetime.now())
-    
-    def __str__(self):
-        return self.nombre
