+++ lab06/lab06/asgi.py
@@ -1,0 +1,16 @@
+"""
+ASGI config for lab06 project.
+
+It exposes the ASGI callable as a module-level variable named ``application``.
+
+For more information on this file, see
+https://docs.djangoproject.com/en/3.2/howto/deployment/asgi/
+"""
+
+import os
+
+from django.core.asgi import get_asgi_application
+
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'lab06.settings')
+
+application = get_asgi_application()
