+++ lab06/tienda/templates/index.html
@@ -1,0 +1,86 @@
+{% extends "layout.html" %} {% block content %}
+<!-- /Block Content -->
+<div
+	id="carouselExampleIndicators"
+	class="carousel slide my-4"
+	data-ride="carousel"
+>
+	<ol class="carousel-indicators">
+		<li
+			data-target="#carouselExampleIndicators"
+			data-slide-to="0"
+			class="active"
+		></li>
+		<li data-target="#carouselExampleIndicators" data-slide-to="1"></li>
+		<li data-target="#carouselExampleIndicators" data-slide-to="2"></li>
+	</ol>
+	<div class="carousel-inner" role="listbox">
+		<div class="carousel-item active">
+			<img
+				class="d-block img-fluid"
+				src="http://placehold.it/900x350"
+				alt="First slide"
+			/>
+		</div>
+		<div class="carousel-item">
+			<img
+				class="d-block img-fluid"
+				src="http://placehold.it/900x350"
+				alt="Second slide"
+			/>
+		</div>
+		<div class="carousel-item">
+			<img
+				class="d-block img-fluid"
+				src="http://placehold.it/900x350"
+				alt="Third slide"
+			/>
+		</div>
+	</div>
+	<a
+		class="carousel-control-prev"
+		href="#carouselExampleIndicators"
+		role="button"
+		data-slide="prev"
+	>
+		<span class="carousel-control-prev-icon" aria-hidden="true"></span>
+		<span class="sr-only">Previous</span>
+	</a>
+	<a
+		class="carousel-control-next"
+		href="#carouselExampleIndicators"
+		role="button"
+		data-slide="next"
+	>
+		<span class="carousel-control-next-icon" aria-hidden="true"></span>
+		<span class="sr-only">Next</span>
+	</a>
+</div>
+
+<div class="row">
+	{% for prod in product_list %}
+
+	<div class="col-lg-4 col-md-6 mb-4">
+		<div class="card h-100">
+			<a href="#"
+				><img class="card-img-top" src="http://placehold.it/700x400" alt=""
+			/></a>
+			<div class="card-body">
+				<h4 class="card-title">
+					<a href="producto/{{prod.id}}">{{prod.nombre}}</a>
+				</h4>
+				<h5>S/. {{prod.precio}}</h5>
+				<p class="card-text">Categoria: {{prod.Categoria.nombre}}</p>
+			</div>
+			<div class="card-footer">
+				<small class="text-muted"
+					>&#9733; &#9733; &#9733; &#9733; &#9734;</small
+				>
+			</div>
+		</div>
+	</div>
+	{% endfor %}
+</div>
+
+<!-- /Block Content -->
+{% endblock %}
