+++ lab06/tienda/templates/layout.html
@@ -1,0 +1,105 @@
+{% load static%}
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<meta charset="utf-8" />
+		<meta
+			name="viewport"
+			content="width=device-width, initial-scale=1, shrink-to-fit=no"
+		/>
+		<meta name="description" content="" />
+		<meta name="author" content="" />
+
+		<title>Tienda</title>
+
+		<!-- Bootstrap core CSS -->
+		<link
+			href="{% static 'vendor/bootstrap/css/bootstrap.min.css' %}"
+			rel="stylesheet"
+		/>
+
+		<!-- Custom styles for this template -->
+		<link href="{% static 'css/shop.css' %}" rel="stylesheet" />
+	</head>
+
+	<body>
+		<!-- Navigation -->
+		<nav class="navbar navbar-expand-lg navbar-dark bg-dark fixed-top">
+			<div class="container">
+				<a class="navbar-brand" href="{% url 'tienda:index' %}">Nyan's Corp</a>
+				<button
+					class="navbar-toggler"
+					type="button"
+					data-toggle="collapse"
+					data-target="#navbarResponsive"
+					aria-controls="navbarResponsive"
+					aria-expanded="false"
+					aria-label="Toggle navigation"
+				>
+					<span class="navbar-toggler-icon"></span>
+				</button>
+				<div class="collapse navbar-collapse" id="navbarResponsive">
+					<ul class="navbar-nav ml-auto">
+						<li class="nav-item active">
+							<a class="nav-link" href="#"
+								>Home
+								<span class="sr-only">(current)</span>
+							</a>
+						</li>
+						<li class="nav-item">
+							<a class="nav-link" href="#">About</a>
+						</li>
+						<li class="nav-item">
+							<a class="nav-link" href="#">Services</a>
+						</li>
+						<li class="nav-item">
+							<a class="nav-link" href="#">Contact</a>
+						</li>
+					</ul>
+				</div>
+			</div>
+		</nav>
+
+		<!-- Page Content -->
+		<div class="container">
+			<div class="row">
+				<div class="col-lg-3">
+					<h1 class="my-4">Nyan's Corp</h1>
+					<div class="list-group">
+						{% for categoria in categorias %}
+						<a
+							href="{% url 'tienda:categoria' categoria.id %}"
+							class="list-group-item"
+						>
+							{{categoria.nombre}}
+						</a>
+						{% endfor %}
+					</div>
+				</div>
+				<!-- /.col-lg-3 -->
+
+				<div class="col-lg-9">
+					{% block content %} {% endblock %}
+					<!-- /.row -->
+				</div>
+				<!-- /.col-lg-9 -->
+			</div>
+			<!-- /.row -->
+		</div>
+		<!-- /.container -->
+
+		<!-- Footer -->
+		<footer class="py-5 bg-dark">
+			<div class="container">
+				<p class="m-0 text-center text-white">
+					Copyright &copy; Your Website 2020
+				</p>
+			</div>
+			<!-- /.container -->
+		</footer>
+
+		<!-- Bootstrap core JavaScript -->
+		<script src="{% static 'vendor/jquery/jquery.min.js' %}"></script>
+		<script src="{% static 'vendor/bootstrap/js/bootstrap.bundle.min.js' %}"></script>
+	</body>
+</html>
