+++ lab06/tienda/templates/producto.html
@@ -1,0 +1,3 @@
+{% extends "layout.html" %} {% block content %}
+<h1>{{producto.nombre}}</h1>
+{% endblock %}
