+++ lab06/tienda/tests.py
